refactor: remove barrel export for history utility
